added checks for display controller and display service

+++ EnergySocialNetwork/src/main/java/com/revature/EnergySocialNetwork/controllers/DisplayController.java
@@ -39,9 +39,17 @@ public class DisplayController {
      * @param profileId users Id given when the profile was made
      * @return get all displays made by one user
      */
-    @GetMapping("profile/{profileId}")
-    public List<Display> getAllByProfileId(@PathVariable Integer profileId) {
-        return displayService.getAllByProfileId(profileId);
+    @GetMapping("profile/{profileId}") //checked
+    public JsonResponse getAllByProfileId(@PathVariable Integer profileId) {
+        JsonResponse jsonResponse;
+        List<Display> displaysFromDb = displayService.getAllByProfileId(profileId);
+        if (displaysFromDb == null){
+            jsonResponse = new JsonResponse(false, "No displays match this profileId", null);
+        } else {
+            jsonResponse = new JsonResponse(true, "Displaying all displays from ProfileId: " + profileId, displaysFromDb);
+        }
+
+        return jsonResponse;
     }
 
     /**
@@ -49,9 +57,16 @@ public class DisplayController {
      * @param displayId Id of the displays made within the database
      * @return getting one singular display by displayId
      */
-    @GetMapping("display/{displayId}")
-    public Display getOneByDisplayId(@PathVariable Integer displayId) {
-        return displayService.getOneByDisplayId(displayId);
+    @GetMapping("display/{displayId}") //checked
+    public JsonResponse getOneByDisplayId(@PathVariable Integer displayId) {
+        JsonResponse jsonResponse;
+        Display displayFromDb = displayService.getOneByDisplayId(displayId);
+        if (displayFromDb == null){
+            jsonResponse = new JsonResponse(false, "There are no displays matching this displayId", null);
+        } else {
+            jsonResponse = new JsonResponse(true, "Displaying display with displayId: " + displayId, displayFromDb);
+        }
+        return jsonResponse;
     }
 
     /**
@@ -59,10 +74,18 @@ public class DisplayController {
      * @param displayId Id of the displays made within the database
      * @return deleting a display by displayId
      */
-    @DeleteMapping("delete/{displayId}")
-    public Display deleteDisplay(@PathVariable Integer displayId) {
-        displayService.deleteDisplay(displayId);
-        return null;
+    @DeleteMapping("delete/{displayId}") //checked
+    public JsonResponse deleteDisplay(@PathVariable Integer displayId) {
+        JsonResponse jsonResponse;
+        Display displayFromDb = displayService.getOneByDisplayId(displayId);
+        if (displayFromDb == null) {
+            jsonResponse = new JsonResponse(false, "There are no displays matching this displayId", null);
+        } else {
+            jsonResponse = new JsonResponse(true, "Display with displayId " + displayId + " has been deleted.", null);
+            displayService.deleteDisplay(displayId);
+        }
+
+        return jsonResponse;
     }
 
     /**
@@ -70,18 +93,28 @@ public class DisplayController {
      * @param display variable for when a display is sent to the backend
      * @return creating a display
      */
-    @PostMapping
-    public Display createDisplay(@RequestBody Display display) {
-        return displayService.createDisplay(display);
+    @PostMapping //checked
+    public JsonResponse createDisplay(@RequestBody Display display) {
+        JsonResponse jsonResponse;
+        Display displayFromDb = displayService.createDisplay(display);
+        if (displayFromDb == null) {
+            jsonResponse = new JsonResponse(false, "Display needs to include image or description", null);
+        } else {
+            jsonResponse = new JsonResponse(true, "Display has been created", displayFromDb);
+        }
+        return jsonResponse;
     }
 
-    @PatchMapping("{displayId}/profile/{profileId}")
-    public Display addLikeToDisplay(@PathVariable Integer profileId,@PathVariable Integer displayId){
+    @PatchMapping("{displayId}/profile/{profileId}") //checked
+    public JsonResponse addLikeToDisplay(@PathVariable Integer profileId,@PathVariable Integer displayId){
+        JsonResponse jsonResponse;
         Display displayFromDB = displayService.addToLike(profileId,displayId);
-        if (displayFromDB != null){
-            return displayFromDB;
+        if (displayFromDB == null){
+            jsonResponse = new JsonResponse(false, "Profile has already liked this post", null);
+        } else {
+            jsonResponse = new JsonResponse(true, "Profile with profileId: " + profileId + "has liked Display with displayId: " + displayId, displayFromDB);
         }
-        return null;
+        return jsonResponse;
     }
 
     @GetMapping("likers/{displayId}")
